Initial Commit:
This project is already started - I didn't realise that I was going to
keep and develop this, hence the late git-ing.
Some of the code has departed a little from the tutorial as I try to
make it a bit 'individual'.

In addintion to the tutorial material, I will be also try to
containerize this app - Docker. This part has as yet not been added.

Changes to be committed:
      new file:   .gitignore
      new file:   README.md
      new file:   app.py
      new file:   requirements.txt
      new file:   static/bootstrap.min.css
      new file:   static/bootstrap.min.css.map
      new file:   static/bootstrap.min.js
      new file:   templates/index.html
      new file:   templates/login.html
      new file:   templates/welcome.html